+++ _test/flutter_test/lib/main.dart
@@ -1,0 +1,10 @@
+import 'package:df_safer_dart/df_safer_dart.dart';
+import 'package:flutter/material.dart';
+
+void main() {
+  debugAssertErr = true;
+  // This should trigger an assert!
+  final err = Err('test');
+  print(err);
+  runApp(const MaterialApp(home: SizedBox.shrink()));
+}