+++ lib/src/err_model/_err_model.g.dart
@@ -24,7 +24,7 @@ class ErrModel {
   final String? type;
 
   /// No description provided.
-  final String? debugPath;
+  final String? location;
 
   /// No description provided.
   final String? error;
@@ -39,7 +39,7 @@ class ErrModel {
   /// from optional and required parameters.
   const ErrModel({
     required this.type,
-    this.debugPath,
+    this.location,
     required this.error,
     this.statusCode,
     this.stackTrace,
@@ -48,14 +48,14 @@ class ErrModel {
   /// Creates a copy of this instance, replacing the specified fields.
   ErrModel copyWith({
     String? type,
-    String? debugPath,
+    String? location,
     String? error,
     int? statusCode,
     List<String>? stackTrace,
   }) {
     return ErrModel(
       type: type ?? this.type,
-      debugPath: debugPath ?? this.debugPath,
+      location: location ?? this.location,
       error: error ?? this.error,
       statusCode: statusCode ?? this.statusCode,
       stackTrace: stackTrace ?? this.stackTrace,
@@ -65,14 +65,14 @@ class ErrModel {
   /// Creates a copy of this instance, removing the specified fields.
   ErrModel copyWithout({
     bool type = true,
-    bool debugPath = true,
+    bool location = true,
     bool error = true,
     bool statusCode = true,
     bool stackTrace = true,
   }) {
     return ErrModel(
       type: type ? this.type : null,
-      debugPath: debugPath ? this.debugPath : null,
+      location: location ? this.location : null,
       error: error ? this.error : null,
       statusCode: statusCode ? this.statusCode : null,
       stackTrace: stackTrace ? this.stackTrace : null,
@@ -85,11 +85,11 @@ class ErrModel {
   @pragma('vm:prefer-inline')
   String get type$ => type!;
 
-  /// Returns the value of the [debugPath] field.
+  /// Returns the value of the [location] field.
   /// If the field is nullable, the return value may be null; otherwise, it
   /// will always return a non-null value.
   @pragma('vm:prefer-inline')
-  String? get debugPath$ => debugPath;
+  String? get location$ => location;
 
   /// Returns the value of the [error] field.
   /// If the field is nullable, the return value may be null; otherwise, it