+++ lib/src/monads/monad.dart
@@ -77,7 +77,7 @@ sealed class Monad<T extends Object> implements Equatable {
     }
   }
 
-  FutureOr<T> unwrap({int stackLevel = 1});
+  FutureOr<T> unwrap({int delta = 1});
 
   FutureOr<T> unwrapOr(T fallback);